+++ hohlova.polina/common/BinarySearchTree.hpp
@@ -452,19 +452,4 @@ typename BinarySearchTree< Key, Value, Compare >::Iterator BinarySearchTree< Key
     return Iterator();
 }
 
-int sum(int a, int b);
-
-struct Key_summ {
-
-    void operator()(const std::pair<const int, std::string>& key_value)
-    {
-        result_ = sum(result_, key_value.first);
-        if (str.empty())
-            str += key_value.second;
-        else
-            str += " " + key_value.second;
-    }
-    int result_ = 0;
-    std::string str;
-};
 #endif
